sidecar/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
supabase==2.2.0
pydantic==2.5.2
python-dotenv==1.0.0
